novas rotas e testes

+++ app/Http/Controllers/TodoController.php
@@ -9,21 +9,26 @@ class TodoController extends Controller
 {
     public function index(Request $request)
     {
-        $model = Todo::all();
+        $model = Todo::paginate();
 
-        return response()->json([$model], 200);
+        return response()->json($model);
     }
 
     public function store(Request $request)
     {
         $model = Todo::create($request->only(['title' , 'description']));
 
-        return response()->json([$model], 201);
+        return response()->json($model, 201);
     }
 
     public function show($id)
     {
         $todo = Todo::find($id);
+
+        if (!$todo)
+        {
+            return response()->json(['error' => 'Not Found'], 404);
+        }
 
         return response()->json($todo);
     }
@@ -31,6 +36,11 @@ class TodoController extends Controller
     public function update(Request $request, $id)
     {
         $todo = Todo::find($id);
+
+        if (!$todo)
+        {
+            return response()->json(['error' => 'Not Found'], 404);
+        }
 
         $todo->update($request->all());
 
@@ -40,9 +50,33 @@ class TodoController extends Controller
     public function destroy($id)
     {
         $todo = Todo::find($id);
+
+        if (!$todo)
+        {
+            return response()->json(['error' => 'Not Found'], 404);
+        }
+
         $todo->delete();
 
         return response()->json(['Todo Deleted'], 204);
+    }
+
+    public function doneTodo(Request $request,$id)
+    {
+        $todo = Todo::find($id);
+
+        if (!$todo)
+        {
+            return response()->json(['error' => 'Not Found'], 404);
+        }
+        if ($todo['done'] == 0)
+        {
+            $todo->done();
+        }else{
+            $todo->undone();
+        }
+
+        return response()->json($todo);
     }
